pyproject.toml
[project]
name = "animo-disconnect-telegram"
version = "0.1.0"
description = ""
authors = [
    {name = "Joshua Armaine Gilo",email = "joshua_gilo@dlsu.edu.ph"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot[job-queue] (>=22.0,<23.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "aiohttp (>=3.11.18,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
